test/requirements.txt
nose-testconfig
unittest2
requests
